requirements.txt
Django==4.2
psycopg2-binary==2.9.3
djangorestframework==3.14.0
django-cors-headers==3.14.0
gunicorn==20.1.0
whitenoise==6.4.0
django-environ==0.8.1
django-allauth
requests